+++ addons/mail/static/tests/thread/unread_messages_banner.test.js
@@ -8,16 +8,13 @@ import {
     start,
     startServer,
 } from "@mail/../tests/mail_test_helpers";
-import { Thread } from "@mail/core/common/thread";
 import { describe, test } from "@odoo/hoot";
-import { delay } from "@odoo/hoot-dom";
 import { tick } from "@odoo/hoot-mock";
 import {
     asyncStep,
     Command,
     getService,
     onRpc,
-    patchWithCleanup,
     serverState,
     waitForSteps,
     withUser,
@@ -67,47 +64,6 @@ test("mark thread as read from unread messages banner", async () => {
         parent: ["span", { text: "30 new messagesMark as Read" }],
     });
     await contains(".o-mail-Thread-jumpToUnread", { count: 0 });
-});
-
-test("scroll to the first unread message (slow ref registration)", async () => {
-    const pyEnv = await startServer();
-    const channelId = pyEnv["discuss.channel"].create({ name: "general" });
-    const messageIds = [];
-    for (let i = 0; i < 201; i++) {
-        messageIds.push(
-            pyEnv["mail.message"].create({
-                body: `message ${i}`,
-                model: "discuss.channel",
-                res_id: channelId,
-            })
-        );
-    }
-    const [selfMember] = pyEnv["discuss.channel.member"].search([
-        ["partner_id", "=", serverState.partnerId],
-        ["channel_id", "=", channelId],
-    ]);
-    pyEnv["discuss.channel.member"].write([selfMember], { new_message_separator: messageIds[100] });
-    let slowRegisterMessageRef = false;
-    patchWithCleanup(Thread.prototype, {
-        async registerMessageRef() {
-            if (slowRegisterMessageRef) {
-                // Ensure scroll is made even when messages are mounted later.
-                await delay(500);
-            }
-            super.registerMessageRef(...arguments);
-        },
-    });
-    await start();
-    await openDiscuss(channelId);
-    await click("[title='Jump to Present']");
-    await scroll(".o-mail-Thread", "bottom");
-    await contains(".o-mail-Thread", { scroll: "bottom" });
-    slowRegisterMessageRef = true;
-    await click("span", {
-        text: "101 new messages",
-        parent: ["span", { text: "101 new messagesMark as Read" }],
-    });
-    await isInViewportOf(".o-mail-Message:contains(message 100)", ".o-mail-Thread");
 });
 
 test("scroll to unread notification", async () => {
